requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
deepface==0.0.93
opencv-python==4.10.0.84

